+++ back/src/main/java/com/example/monday/resource/CreateStudent.java
@@ -24,5 +24,5 @@ public class CreateStudent {
     private Kierunek kierunek;
 
     @NotNull
-    private Integer ects;
+    private Long ects;
 }
